+++ .env
@@ -1,0 +1,2 @@
+DB_HOST="mongodb://localhost:27017/urlMoniture"
+CHECKINTERVAL="5"
