+++ configurations/cronJobConfig.js
@@ -1,13 +1,13 @@
 import nodeCron from "node-cron";
-import { createReport } from "../shared/createReport.js";
+import { creatReport } from "../shared/createReport.js";
 
 const cronJob = nodeCron.schedule(
   //   "0 */5 * * * *",
-  "*/55 * * * * *",
-//   "*/1000 * * * * *",
-  function job() {
+  "*/5 * * * * *",
+  //   "*/1000 * * * * *",
+  async function job() {
     console.log(new Date().toLocaleString());
-    createReport();
+    await creatReport();
   },
   { scheduled: false }
 );
